+++ contracts/cw-rules/src/tests/daodao.rs
@@ -1,8 +1,8 @@
-use cosmwasm_std::{to_binary, Addr, Binary, Empty, Uint128};
+use cosmwasm_std::{to_binary, Addr, Binary, Uint128};
 use cw20::Cw20Coin;
 use cw20_staked_balance_voting::msg::ActiveThreshold;
 use cw_core::state::ProposalModule;
-use cw_multi_test::{next_block, App, Contract, ContractWrapper, Executor};
+use cw_multi_test::{next_block, App, Executor};
 use cw_proposal_multiple::{
     state::{MultipleChoiceOption, MultipleChoiceOptions},
     voting_strategy::VotingStrategy,
@@ -15,76 +15,13 @@ use voting::{
 
 use cw_rules_core::msg::{InstantiateMsg, QueryMsg, RuleResponse};
 
-const CREATOR_ADDR: &str = "creator";
+use crate::tests::helpers::{
+    cw_rules_contract, multiple_proposal_contract, single_proposal_contract,
+};
 
-fn cw_rules_contract() -> Box<dyn Contract<Empty>> {
-    let contract = ContractWrapper::new(
-        crate::contract::execute,
-        crate::contract::instantiate,
-        crate::contract::query,
-    );
-    Box::new(contract)
-}
-
-fn cw20_contract() -> Box<dyn Contract<Empty>> {
-    let contract = ContractWrapper::new(
-        cw20_base::contract::execute,
-        cw20_base::contract::instantiate,
-        cw20_base::contract::query,
-    );
-    Box::new(contract)
-}
-
-fn cw20_stake_contract() -> Box<dyn Contract<Empty>> {
-    let contract = ContractWrapper::new(
-        cw20_stake::contract::execute,
-        cw20_stake::contract::instantiate,
-        cw20_stake::contract::query,
-    );
-    Box::new(contract)
-}
-
-fn single_proposal_contract() -> Box<dyn Contract<Empty>> {
-    let contract = ContractWrapper::new(
-        cw_proposal_single::contract::execute,
-        cw_proposal_single::contract::instantiate,
-        cw_proposal_single::contract::query,
-    )
-    .with_reply(cw_proposal_single::contract::reply)
-    .with_migrate(cw_proposal_single::contract::migrate);
-    Box::new(contract)
-}
-
-fn multiple_proposal_contract() -> Box<dyn Contract<Empty>> {
-    let contract = ContractWrapper::new(
-        cw_proposal_multiple::contract::execute,
-        cw_proposal_multiple::contract::instantiate,
-        cw_proposal_multiple::contract::query,
-    )
-    .with_reply(cw_proposal_multiple::contract::reply)
-    .with_migrate(cw_proposal_multiple::contract::migrate);
-    Box::new(contract)
-}
-
-fn cw_gov_contract() -> Box<dyn Contract<Empty>> {
-    let contract = ContractWrapper::new(
-        cw_core::contract::execute,
-        cw_core::contract::instantiate,
-        cw_core::contract::query,
-    )
-    .with_reply(cw_core::contract::reply);
-    Box::new(contract)
-}
-
-fn cw20_staked_balances_voting() -> Box<dyn Contract<Empty>> {
-    let contract = ContractWrapper::new(
-        cw20_staked_balance_voting::contract::execute,
-        cw20_staked_balance_voting::contract::instantiate,
-        cw20_staked_balance_voting::contract::query,
-    )
-    .with_reply(cw20_staked_balance_voting::contract::reply);
-    Box::new(contract)
-}
+use super::helpers::{
+    cw20_stake_contract, cw20_staked_balances_voting, cw20_template, cw_gov_contract, CREATOR_ADDR,
+};
 
 fn instantiate_with_staking_active_threshold(
     app: &mut App,
@@ -93,7 +30,7 @@ fn instantiate_with_staking_active_threshold(
     initial_balances: Option<Vec<Cw20Coin>>,
     active_threshold: Option<ActiveThreshold>,
 ) -> Addr {
-    let cw20_id = app.store_code(cw20_contract());
+    let cw20_id = app.store_code(cw20_template());
     let cw20_staking_id = app.store_code(cw20_stake_contract());
     let governance_id = app.store_code(cw_gov_contract());
     let votemod_id = app.store_code(cw20_staked_balances_voting());
