+++ contracts/cw-rules/src/tests/mod.rs
@@ -1,4 +1,5 @@
 mod balance;
 mod daodao;
 mod generic;
+mod helpers;
 mod nft;
